Organize code in folders and create files for tetsing
