+++ app-management/ui_class.rb
