+++ classes/book_class.rb
@@ -21,3 +21,13 @@ class Book
     }
   end
 end
+
+Book
+Person
+Student
+Teacher
+Classroom
+Rental
+Capitalize decorator 
+8. Decorator
+9. Trimmer decorator
