+++ data-management/data_manager_class.rb
