+++ main.rb
@@ -2,7 +2,6 @@ require_relative 'app'
 
 def main
   app = App.new
-  # app.load_data
   app.run
 end
